requirements.txt
astroid==1.5.3
autopep8==1.3.3
dj-database-url==0.4.1
Django==1.11.1
djangorestframework==3.7.3
gunicorn==19.6.0
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.3.2
pycodestyle==2.3.1
pylint==1.7.4
pytz==2017.3
six==1.11.0
whitenoise==3.2
wrapt==1.10.11
